build(npm): use npm workspaces
Rather than maintaining a separate project for the tests, we can use npm workspaces to handle it in
one go.

This feature does mean updating the npm version for the Node 14 test run, as it only has npm 6
available.

The Parcel cache path has to be set for the test project as it ends up in the root otherwise.